[dxvk] Rename descriptor_heap file -> descriptor_info
Need to separate this to avoid circular dependency jank.
